+++ packages/apps/chainweb-evm-dashboard/src/app/chain-comparison/page.tsx
@@ -1,0 +1,5 @@
+export default async function Page() {
+  return (
+    <div>Chain Comparison</div>
+  )
+}