+++ packages/libs/wallet-adapter-walletconnect/src/index.ts
@@ -1,3 +1,4 @@
+export { WALLET_CONNECT_ADAPTER } from './constants';
 export { detectWalletConnectProvider } from './provider';
 export { WalletConnectAdapter } from './WalletConnectAdapter';
 export { createWalletConnectAdapter } from './walletConnectAdapterFactory';