+++ packages/libs/wallet-adapter-zelcore/src/index.ts
@@ -1,3 +1,4 @@
+export { ZELCORE_ADAPTER } from './constants';
 export {
   ZelcoreAdapter,
   createZelcoreAdapter,