+++ packages/apps/wallet-sdk-example/CHANGELOG.md
@@ -1,5 +1,19 @@
 # wallet-sdk-example-app
 
+## 0.0.11
+
+### Patch Changes
+
+- @kadena/client\@1.18.2
+- @kadena/wallet-sdk\@0.2.5
+
+## 0.0.10
+
+### Patch Changes
+
+- @kadena/client\@1.18.1
+- @kadena/wallet-sdk\@0.2.4
+
 ## 0.0.9
 
 ### Patch Changes