+++ packages/libs/client-examples/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @kadena/client-examples
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies \[c21ada4]
+  - @kadena/chainweb-node-client\@0.9.4
+  - @kadena/client\@1.18.2
+  - @kadena/client-utils\@0.13.3
+
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies \[c967a57]
+  - @kadena/pactjs\@0.5.0
+  - @kadena/chainweb-node-client\@0.9.2
+  - @kadena/client\@1.18.1
+  - @kadena/client-utils\@0.13.2
+
 ## 0.1.15
 
 ### Patch Changes