+++ packages/libs/pactjs-generator/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log - @kadena/pactjs-generator
 
+## 1.18.2
+
+## 1.18.1
+
 ## 1.18.0
 
 ## 1.17.1