+++ packages/libs/wallet-adapter-magic/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @kadena/wallet-adapter-magic
 
+## 0.0.6
+
+### Patch Changes
+
+- @kadena/client\@1.18.2
+- @kadena/wallet-adapter-core\@0.1.2
+
+## 0.0.5
+
+### Patch Changes
+
+- @kadena/client\@1.18.1
+- @kadena/wallet-adapter-core\@0.1.1
+
 ## 0.0.4
 
 ### Patch Changes