+++ packages/libs/wallet-adapter-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @kadena/wallet-adapter-react
 
+## 0.0.6
+
+### Patch Changes
+
+- @kadena/wallet-adapter-core\@0.1.2
+
+## 0.0.5
+
+### Patch Changes
+
+- @kadena/wallet-adapter-core\@0.1.1
+
 ## 0.0.4
 
 ### Patch Changes