+++ packages/tools/kadena-cli/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @kadena/kadena-cli
 
+## 1.4.5
+
+### Patch Changes
+
+- @kadena/client\@1.18.2
+- @kadena/client-utils\@0.13.3
+- @kadena/pactjs-cli\@1.18.2
+- @kadena/pactjs-generator\@1.18.2
+
+## 1.4.4
+
+### Patch Changes
+
+- Updated dependencies \[c967a57]
+  - @kadena/pactjs\@0.5.0
+  - @kadena/client\@1.18.1
+  - @kadena/client-utils\@0.13.2
+  - @kadena/pactjs-cli\@1.18.1
+  - @kadena/pactjs-generator\@1.18.1
+
 ## 1.4.3
 
 ### Patch Changes